+++ app/src/main/java/com/developbharat/facegate/domain/models/AttendanceSheetRecord.kt
@@ -1,8 +1,8 @@
 package com.developbharat.facegate.domain.models
 
 data class AttendanceSheetRecord(
-    val personId: Long,
-    val batchId: Long,
+    val personId: Int,
+    val batchId: Int,
     val matchScore: Double,
     val frameBase64: String? = null,
 )
